fix tests

+++ boundlessconnect/tests/testerplugin.py
@@ -275,7 +275,8 @@ class BoundlessConnectTests(unittest.TestCase):
 
         self.assertTrue('http://dummyurl.com', settings.value('repoUrl', '', str))
         settings.setValue('repoUrl', oldRepoUrl)
-        os.remove(fName)
+        if os.path.isfile(fName):
+            os.remove(fName)
 
     @classmethod
     def tearDownClass(cls):
